Consolidated common koin modules to one file (variants left alone for now)
