--- app/src/debug/java/com/atomicrobot/carbon/modules/CrashReporterModule.kt
@@ -1,10 +1,0 @@
-package com.atomicrobot.carbon.modules
-
-import com.atomicrobot.carbon.monitoring.LoggingOnlyCrashReporter
-import org.koin.dsl.module
-
-val crashReporterModule = module {
-    single {
-        LoggingOnlyCrashReporter()
-    }
-}
